fixed bugs, including not working white/blacklist
- not whitelisted tags where not deleted at import, making
  delete-existing useless. Deletion of those can be controlled
  with clear_blacklisted, which is set in mmusicc with the
  delete existing value. Also updated test to cover this case
- improved help message and renamed --lazy to --lazy-import
  --lazy will still work since is a abbreviation
- added Opus codec to xiph
- fixed bug in mp3 where under some circumstances a tag would
  be overlooked. Either a warning is writen to log or the tag
  is deleted when delete_existing is True.
- fixed bug of logging module with pytest

+++ mmusicc/util/ffmpeg.py
@@ -44,7 +44,8 @@ class FFmpeg(object):
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         if not self.exit_status == 0:
-            self._target.unlink()
+            if self._target.exists():
+                self._target.unlink()
 
     def run(self):
         """Execute ffmpeg command line. Log stderr output.
